requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
html==1.16
lxml==3.6.4
pudb==2016.2
Pygments==2.1.3
python-docx==0.8.6
requests==2.11.1
urwid==1.3.1
